serverless/requirements.txt
requests==2.31.0
boto3==1.28.62
firebase-admin==6.2.0
psycopg2-binary==2.9.7
python-decouple==3.8
